telegram_service/requirements.txt
python-telegram-bot>=22.0
fastapi>=0.100.0
uvicorn>=0.23.0
httpx>=0.25.0
